requirements.txt
Django==4.2.5
psycopg2-binary>=2.8
pandas==2.1.0
djangorestframework==3.14.0
scikit-learn==1.3
xgboost==2.0.0
lightgbm==4.1.0